Move sort operation to separate file
